backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
langchain-google-genai
langchain
requests
beautifulsoup4
